Update: User를 얻어오는 로직 변경
- AuthenticationPrincipal 어노테이션으로 UserDetails를 받아오도록 수정
- User정보를 가져오는 로직을 Service에서 Repository로 이전
- SecurityUtil 삭제
