--- src/main/java/com/divide/security/SecurityUtil.java
@@ -1,32 +1,0 @@
-package com.divide.security;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Optional;
-
-@Slf4j
-public class SecurityUtil {
-    public static Optional<String> getCurrentEmail() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null) {
-            log.debug("Security Context에 인증 정보가 없습니다.");
-            return Optional.empty();
-        }
-
-        String email = null;
-        if (authentication.getPrincipal() instanceof UserDetails) {
-            UserDetails springSecurityUser = (UserDetails) authentication.getPrincipal();
-            email = springSecurityUser.getUsername();
-        } else if (authentication.getPrincipal() instanceof String) {
-            email = (String) authentication.getPrincipal();
-        } else {
-            log.error("authentication의 principal의 type이 " + authentication.getPrincipal().getClass() + " 입니다.");
-        }
-
-        return Optional.ofNullable(email);
-    }
-}
