+++ src/test/java/com/divide/repository/UserServiceTest.java
@@ -1,9 +1,9 @@
 package com.divide.repository;
 
 import com.divide.BaseException;
+import com.divide.user.UserService;
 import com.divide.user.dto.request.SignupRequest;
 import com.divide.user.UserRepository;
-import com.divide.user.UserService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
